% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos_ggplot.R
\name{nber_load_ggplot}
\alias{nber_load_ggplot}
\title{Loads the dates of begining and ending of recessions acording to NBER to be used with ggplot2}
\usage{
nber_load_ggplot(periodicity)
}
\arguments{
\item{periodicity}{What is the periodicity of data used. Only accepts (q)uartely or (m)onthly. You can use the whole word or only the first letter}
}
\value{
An matrix with two columns: the first is the beggining of the recession according to NBER and the second column the end
}
\description{
Loads the dates of begining and ending of recessions acording to NBER to be used with ggplot2
}
\note{
The diference between this command and \code{\link{nber_load}} is that this one uses dates in the usual way (human readable), while \code{\link{nber_load}} uses the \code{ts} convention, i.e. year.fraction
}
\seealso{
\code{\link{nber_load}}
}
