% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos_ggplot.R
\name{nber_recession_ggplot}
\alias{nber_recession_ggplot}
\title{A convenient wrapper to plot shadded areas in ggplot2 in periods of recession}
\usage{
nber_recession_ggplot(periodicity, start_date, end_date, alpha = 0.3)
}
\arguments{
\item{periodicity}{What is the periodicity of data used. Only accepts (q)uartely or (m)onthly. You can use the whole word or only the first letter}

\item{start_date}{The start date in \code{\link{Dates}} format}

\item{end_date}{The ending date in \code{\link{Dates}} format}

\item{alpha}{The transparency. Defaults to 0.3}
}
\value{
This should be used with a ggplot. Will shade the areas where there were recessions
}
\description{
A convenient wrapper to plot shadded areas in ggplot2 in periods of recession
}
\note{
This basically call \code{\link{geom_rect}} from ggplot2 using data that was loaded and cutted with \code{\link{recession_dates_ggplot}}
}
\examples{
x <- rnorm(500)
today <- Sys.Date()
datess <- rev(seq(today,length.out = 500, by = "-1 month"))

ddata <- data.frame(Dates = datess, x = x)

ggplot() + geom_line(data = ddata, aes(Dates,x)) + nber_recession_ggplot("m",start_date = ddata[1,1],end_date =  ddata[500,1])

}
\seealso{
\code{\link{recession_dates_ggplot}}
}
