% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos.R
\name{plot_recession}
\alias{plot_recession}
\title{Plot using base plot}
\usage{
plot_recession(periodicity, ytop = 700, ybottom = -700, alpha = 0.3)
}
\arguments{
\item{periodicity}{What is the periodicity of data used. Only accepts (q)uartely or (m)onthly. You can use the whole word or only the first letter}

\item{ytop, ybottom}{Controls the upper and lower limits of the shading. The default should be enough for most data}

\item{alpha}{Controls the transparency of the shade. Should be a number between 0 and 1}
}
\value{
If there is a time series already plot, it should shade the back. Otherwise, it returns an error.
}
\description{
Plot using base plot
}
\examples{
x <- rnorm(100)
x <- ts(x,start = c(1900,01),freq=4)
plot(x)
plot_recession("q")
}
