% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos_ggplot.R
\name{recession_dates_ggplot}
\alias{recession_dates_ggplot}
\title{Cuts to get the dates of recessions between two dates}
\usage{
recession_dates_ggplot(periodicity, start_date, end_date)
}
\arguments{
\item{periodicity}{What is the periodicity of data used. Only accepts (q)uartely or (m)onthly. You can use the whole word or only the first letter}

\item{start_date}{The start date in \code{\link{Dates}} format}

\item{end_date}{The ending date in \code{\link{Dates}} format}
}
\value{
A dataframe with two columns, the first with the start date of all recessions between \code{start_date} and \code{end_date} and the second column with the end date of all recessions between \code{start_date} and \code{end_date}
}
\description{
Cuts to get the dates of recessions between two dates
}
