% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recession_q_ggplot}
\alias{recession_q_ggplot}
\title{Dates of recessions in USA, according to NBER (quartely)}
\format{This data "human friendly", so it is actually a string of data in the format YYYY-MM-DD, and it is used by ggplot. This is not true to the data used in the base plot.Each dataset contains two columns:
\describe{
\item{start_date}{The quarter of the begin of a recession}
\item{end_date}{The quarter of the end of a recession}
}}
\source{
\url{https://fred.stlouisfed.org/series/USREC}
}
\usage{
recession_q_ggplot
}
\description{
Dates of recessions in USA, according to NBER (quartely)
}
\keyword{datasets}
