% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recessions_quartely}
\alias{recessions_quartely}
\title{Dates of recessions in USA, according to NBER (quartely)}
\format{This data _is not_ "human friendly", it is a time in the format YYYY.Frac, which Frac is a fraction of the year. This is what you get if you take an \code{ts} object\code{x} and use \code{time(x)}. This is use by the base plot function. Each dataset contains two columns:
\describe{
\item{start_date}{The quarter of the beginning of a recession}
\item{end_date}{The quarter of the end of a recession}
}}
\source{
\url{https://fred.stlouisfed.org/series/USREC}
}
\usage{
recessions_quartely
}
\description{
Dates of recessions in USA, according to NBER (quartely)
}
\keyword{datasets}
