% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foos.R
\name{nber_load}
\alias{nber_load}
\title{Loads the dates of begining and ending of recessions acording to NBER}
\usage{
nber_load(periodicity)
}
\arguments{
\item{periodicity}{What is the periodicity of data used. Only accepts (q)uartely or (m)onthly. You can use the whole word or only the first letter}
}
\value{
An matrix with two columns: the first is the beggining of the recession according to NBER and the second column the end
}
\description{
Loads the dates of begining and ending of recessions acording to NBER
}
\note{
The diference between this command and \code{\link{nber_load_ggplot}} is that this one \code{ts} convention, i.e. year.fraction, while \code{\link{nber_load_ggplot}} uses the data that are human readable
}
\seealso{
\code{\link{nber_load_ggplot}}
}
